lora-mp/pyproject.toml
[project]
name = "lora-mp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "uflash>=2.0.0",
    "pyserial>=3.5",
]
